refactor(gateway): extract session hub telemetry
